requirements.txt
pyyaml
python-dotenv
requests
